Remove genezio credentials from examples

+++ javascript/getting-started/server/task.js
@@ -1,7 +1,10 @@
-import { mongoose } from "mongoose"
-import { MONGO_DB_URI } from "./helper"
-import { TaskModel } from "./models/task"
+import { mongoose } from "mongoose";
+import { TaskModel } from "./models/task";
 import { GenezioDeploy } from "@genezio/types";
+
+const red_color = "\x1b[31m%s\x1b[0m";
+const missing_env_error =
+  "ERROR: Your MONGO_DB_URI environment variable is not properly set, go to https://genez.io/blog/how-to-add-a-mongodb-to-your-genezio-project/ to learn how to integrate your project with Mongo DB";
 
 /**
  * The Task server class that will be deployed on the genezio infrastructure.
@@ -9,7 +12,7 @@ import { GenezioDeploy } from "@genezio/types";
 @GenezioDeploy()
 export class TaskService {
   constructor() {
-    mongoose.set('strictQuery', true);
+    mongoose.set("strictQuery", true);
     this.#connect();
   }
 
@@ -17,7 +20,14 @@ export class TaskService {
    * Private method used to connect to the DB.
    */
   #connect() {
-    mongoose.connect(MONGO_DB_URI);
+    if (!process.env.MONGO_DB_URI) {
+      console.log(red_color, missing_env_error);
+      return;
+    }
+    mongoose.connect(process.env.MONGO_DB_URI || "").catch((err) => {
+      console.log(err);
+      throw err;
+    });
   }
 
   /**
@@ -31,52 +41,67 @@ export class TaskService {
    * @returns An object containing two properties: { success: true, tasks: tasks }
    */
   async getAllTasksByUser(token) {
-    console.log(`Get all tasks by user request received with token ${token}`)
+    if (!process.env.MONGO_DB_URI) {
+      console.log(red_color, missing_env_error);
+      return { success: false, err: missing_env_error };
+    }
+    console.log(`Get all tasks by user request received with token ${token}`);
 
-    const tasks = (await TaskModel.find({ token: token })).map((task) => {
-      return {
-        id: task._id.toString(),
-        title: task.title,
-        solved: task.solved,
-        token: task.token
-        };
-    });
-
-    if (tasks.length === 0) {
-      await TaskModel.create({
-        token: token,
-        title: "Check the other example projects",
-        url: "https://github.com/Genez-io/genezio-examples"
-      })
-
-      await TaskModel.create({
-        token: token,
-        title: "Check our documentation",
-        url: "https://docs.genez.io/genezio-documentation/"
-      })
-
-      await TaskModel.create({
-        token: token,
-        title: "Watch our Youtube tutorials",
-        url: "https://www.youtube.com/@genezio7235"
-      })
-
-      await TaskModel.create({
-        token: token,
-        title: "Read our technical articles on genezio blog",
-        url: "https://genez.io/blog/"
-      })
-
-      const initTasks = (await TaskModel.find({ token: token })).map((task) => {
+    let tasks;
+    try {
+      tasks = (await TaskModel.find({ token: token })).map((task) => {
         return {
           id: task._id.toString(),
           title: task.title,
           solved: task.solved,
-          token: task.token
-          };
+          token: task.token,
+        };
       });
+    } catch (error) {
+      return { success: false, err: error.toString() };
+    }
 
-      return { success: true, tasks: initTasks };
+    if (tasks.length === 0) {
+      try {
+        await TaskModel.create({
+          token: token,
+          title: "Check the other example projects",
+          url: "https://github.com/Genez-io/genezio-examples",
+        });
+
+        await TaskModel.create({
+          token: token,
+          title: "Check our documentation",
+          url: "https://docs.genez.io/genezio-documentation/",
+        });
+
+        await TaskModel.create({
+          token: token,
+          title: "Watch our Youtube tutorials",
+          url: "https://www.youtube.com/@genezio7235",
+        });
+
+        await TaskModel.create({
+          token: token,
+          title: "Read our technical articles on genezio blog",
+          url: "https://genez.io/blog/",
+        });
+
+        const initTasks = (await TaskModel.find({ token: token })).map(
+          (task) => {
+            return {
+              id: task._id.toString(),
+              title: task.title,
+              solved: task.solved,
+              token: task.token,
+            };
+          }
+        );
+
+        return { success: true, tasks: initTasks };
+      } catch (error) {
+        return { success: false, err: err.toString() };
+      }
     }
 
     return { success: true, tasks: tasks };
@@ -94,16 +119,27 @@ export class TaskService {
    * @returns An object containing two properties: { success: true, tasks: tasks }
    */
   async createTask(token, title) {
-    console.log(`Create task request received for user with ${token} with title ${title}`)
+    if (!process.env.MONGO_DB_URI) {
+      console.log(red_color, missing_env_error);
+      return { success: false, err: missing_env_error };
+    }
+    console.log(
+      `Create task request received for user with ${token} with title ${title}`
+    );
 
-    const task = await TaskModel.create({
-      title: title,
-      token: token
-    });
+    let task;
+    try {
+      task = await TaskModel.create({
+        title: title,
+        token: token,
+      });
+    } catch (error) {
+      return { success: false, err: error.toString() };
+    }
 
     return {
       success: true,
-      task: { title: title, token: token.toString(), id: task._id.toString() }
+      task: { title: title, token: token.toString(), id: task._id.toString() },
     };
   }
 
@@ -120,15 +156,25 @@ export class TaskService {
    * @returns An object containing two properties: { success: true }
    */
   async updateTask(token, id, title, solved) {
-    console.log(`Update task request received with id ${id} with title ${title} and solved value ${solved}`)
-
-    await TaskModel.updateOne(
-      { _id: id, token: token },
-      {
-        title: title,
-        solved: solved
-      }
+    if (!process.env.MONGO_DB_URI) {
+      console.log(red_color, missing_env_error);
+      return { success: false, err: missing_env_error };
+    }
+    console.log(
+      `Update task request received with id ${id} with title ${title} and solved value ${solved}`
     );
+
+    try {
+      await TaskModel.updateOne(
+        { _id: id, token: token },
+        {
+          title: title,
+          solved: solved,
+        }
+      );
+    } catch (error) {
+      return { success: false, err: error.toString() };
+    }
 
     return { success: true };
   }
@@ -145,9 +191,17 @@ export class TaskService {
    * @returns An object containing one property: { success: true }
    */
   async deleteTask(token, id) {
-    console.log(`Delete task with id ${id} request received`)
+    if (!process.env.MONGO_DB_URI) {
+      console.log(red_color, missing_env_error);
+      return { success: false, err: missing_env_error };
+    }
+    console.log(`Delete task with id ${id} request received`);
 
-    await TaskModel.deleteOne({ token: token, _id: id });
+    try {
+      await TaskModel.deleteOne({ token: token, _id: id });
+    } catch (error) {
+      return { success: false, err: error.toString() };
+    }
 
     return { success: true };
   }
