+++ typescript/getting-started/server/task.ts
@@ -1,38 +1,45 @@
-import mongoose from "mongoose"
-import { MONGO_DB_URI } from "./helper"
-import { TaskModel } from "./models/task"
-import { GenezioDeploy } from "@genezio/types"
+import mongoose from "mongoose";
+import { TaskModel } from "./models/task";
+import { GenezioDeploy } from "@genezio/types";
+
+const red_color = "\x1b[31m%s\x1b[0m";
+const missing_env_error =
+  "ERROR: Your MONGO_DB_URI environment variable is not properly set, go to https://genez.io/blog/how-to-add-a-mongodb-to-your-genezio-project/ to learn how to integrate your project with Mongo DB";
 
 export type Task = {
-  id: string,
-  token: string,
-  title: string,
-  solved: boolean,
-  date: Date,
-}
+  id: string;
+  token: string;
+  title: string;
+  solved: boolean;
+  date: Date;
+};
 
 export type GetTasksResponse = {
-  success: boolean,
-  tasks: Task[]
-}
+  success: boolean;
+  tasks: Task[];
+  err?: string;
+};
 
 export type GetTaskResponse = {
-  success: boolean,
-  task?: Task
-}
+  success: boolean;
+  task?: Task;
+  err?: string;
+};
 
 export type UpdateTaskResponse = {
-  success: boolean,
-}
+  success: boolean;
+  err?: string;
+};
 
 export type DeleteTaskResponse = {
-  success: boolean,
-}
+  success: boolean;
+  err?: string;
+};
 
 /**
  * The Task server class that will be deployed on the genezio infrastructure.
  */
- @GenezioDeploy({type: "jsonrpc"})
+@GenezioDeploy({ type: "jsonrpc" })
 export class TaskService {
   constructor() {
     this.#connect();
@@ -42,7 +49,14 @@ export class TaskService {
    * Private method used to connect to the DB.
    */
   #connect() {
-    mongoose.connect(MONGO_DB_URI);
+    if (!process.env.MONGO_DB_URI) {
+      console.log(red_color, missing_env_error);
+      return;
+    }
+    mongoose.connect(process.env.MONGO_DB_URI || "").catch((err) => {
+      console.log(err);
+      throw err;
+    });
   }
 
   /**
@@ -55,53 +69,95 @@ export class TaskService {
    * @returns An object containing two properties: { success: true, tasks: tasks }
    */
   async getAllTasksByUser(token: string): Promise<GetTasksResponse> {
-    console.log(`Get all tasks by user request received with token ${token}`)
+    if (!process.env.MONGO_DB_URI) {
+      console.log(red_color, missing_env_error);
+      return { success: false, tasks: [], err: missing_env_error };
+    }
+    console.log(`Get all tasks by user request received with token ${token}`);
 
-    const tasks = (await TaskModel.find({ token: token })).map((task) => {
+    let tasks = [];
+    try {
+      tasks = await TaskModel.find({ token: token });
+    } catch (err: any) {
+      return { success: false, tasks: [], err: err.toString() };
+    }
+
+    tasks.map((task: any) => {
       return {
         id: task._id.toString(),
         token: task.token,
         title: task.title,
         solved: task.solved,
-        date: task.date
+        date: task.date,
       };
     });
 
     if (tasks.length === 0) {
-      await TaskModel.create({
-        token: token,
-        title: "Check the other example projects",
-        url: "https://github.com/Genez-io/genezio-examples"
-      })
-
-      await TaskModel.create({
-        token: token,
-        title: "Check our documentation",
-        url: "https://docs.genez.io/genezio-documentation/"
-      })
-
-      await TaskModel.create({
-        token: token,
-        title: "Watch our Youtube tutorials",
-        url: "https://www.youtube.com/@genezio7235"
-      })
-
-      await TaskModel.create({
-        token: token,
-        title: "Read our technical articles on genezio blog",
-        url: "https://genez.io/blog/"
-      })
-
-      const initTasks = (await TaskModel.find({ token: token })).map((task) => {
-        return {
-          id: task._id.toString(),
-          token: task.token,
-          title: task.title,
-          solved: task.solved,
-          date: task.date
-          };
+      try {
+        await TaskModel.create({
+          token: token,
+          title: "Check the other example projects",
+          url: "https://github.com/Genez-io/genezio-examples",
         });
+      } catch (err: any) {
+        return { success: false, tasks: [], err: err.toString() };
+      }
 
+      try {
+        await TaskModel.create({
+          token: token,
+          title: "Check our documentation",
+          url: "https://docs.genez.io/genezio-documentation/",
+        });
+      } catch (err: any) {
+        return { success: false, tasks: [], err: err.toString() };
+      }
+      try {
+        await TaskModel.create({
+          token: token,
+          title: "Check our documentation",
+          url: "https://docs.genez.io/genezio-documentation/",
+        });
+      } catch (err: any) {
+        return { success: false, tasks: [], err: err.toString() };
+      }
+
+      try {
+        await TaskModel.create({
+          token: token,
+          title: "Watch our Youtube tutorials",
+          url: "https://www.youtube.com/@genezio7235",
+        });
+      } catch (err: any) {
+        return { success: false, tasks: [], err: err.toString() };
+      }
+
+      try {
+        await TaskModel.create({
+          token: token,
+          title: "Read our technical articles on genezio blog",
+          url: "https://genez.io/blog/",
+        });
+      } catch (err: any) {
+        return { success: false, tasks: [], err: err.toString() };
+      }
+
+      let initTasks = [];
+      try {
+        initTasks = (await TaskModel.find({ token: token })).map(
+          (task: any) => {
+            return {
+              id: task._id.toString(),
+              token: task.token,
+              title: task.title,
+              solved: task.solved,
+              date: task.date,
+            };
+          }
+        );
+      } catch (err: any) {
+        return { success: false, tasks: [], err: err.toString() };
+      }
 
       return { success: true, tasks: initTasks };
     }
@@ -120,16 +176,32 @@ export class TaskService {
    * @returns An object containing two properties: { success: true, tasks: tasks }
    */
   async createTask(token: string, title: string): Promise<GetTaskResponse> {
-    console.log(`Create task request received for user with ${token} with title ${title}`)
+    if (!process.env.MONGO_DB_URI) {
+      console.log(red_color, missing_env_error);
+      return { success: false, err: missing_env_error };
+    }
+    console.log(
+      `Create task request received for user with ${token} with title ${title}`
+    );
 
-    const task = await TaskModel.create({
-      title: title,
-      token: token
-    });
-
+    let task;
+    try {
+      task = await TaskModel.create({
+        title: title,
+        token: token,
+      });
+    } catch (err: any) {
+      return { success: false, err: err.toString() };
+    }
     return {
       success: true,
-      task: { title: title, token:token, id: task._id.toString(), solved: false, date: new Date() }
+      task: {
+        title: title,
+        token: token,
+        id: task._id.toString(),
+        solved: false,
+        date: new Date(),
+      },
     };
   }
 
@@ -145,16 +217,31 @@ export class TaskService {
    * @param {*} solved If the task is solved or not.
    * @returns An object containing two properties: { success: true }
    */
-  async updateTask(token: string, id: string, title: string, solved: boolean) : Promise<UpdateTaskResponse>{
-    console.log(`Update task request received with id ${id} with title ${title} and solved value ${solved}`)
-
-    await TaskModel.updateOne(
-      { _id: id, token: token },
-      {
-        title: title,
-        solved: solved
-      }
+  async updateTask(
+    token: string,
+    id: string,
+    title: string,
+    solved: boolean
+  ): Promise<UpdateTaskResponse> {
+    if (!process.env.MONGO_DB_URI) {
+      console.log(red_color, missing_env_error);
+      return { success: false, err: missing_env_error };
+    }
+    console.log(
+      `Update task request received with id ${id} with title ${title} and solved value ${solved}`
     );
+
+    try {
+      await TaskModel.updateOne(
+        { _id: id, token: token },
+        {
+          title: title,
+          solved: solved,
+        }
+      );
+    } catch (err: any) {
+      return { success: false, err: err.toString() };
+    }
 
     return { success: true };
   }
@@ -169,10 +256,18 @@ export class TaskService {
    * @param {*} id The task's id.
    * @returns An object containing one property: { success: true }
    */
-  async deleteTask(token: string, id: string):Promise<DeleteTaskResponse> {
-    console.log(`Delete task with id ${id} request received`)
+  async deleteTask(token: string, id: string): Promise<DeleteTaskResponse> {
+    if (!process.env.MONGO_DB_URI) {
+      console.log(red_color, missing_env_error);
+      return { success: false, err: missing_env_error };
+    }
+    console.log(`Delete task with id ${id} request received`);
 
-    await TaskModel.deleteOne({ token: token, _id: id });
+    try {
+      await TaskModel.deleteOne({ token: token, _id: id });
+    } catch (err: any) {
+      return { success: false, err: err.toString() };
+    }
 
     return { success: true };
   }
