+++ typescript/todo-list-flutter/server/task.ts
@@ -1,35 +1,42 @@
-import  mongoose from "mongoose"
-import { ObjectId } from 'mongodb'
-import { MONGO_DB_URI } from "./helper"
-import { TaskModel } from "./models/task"
-import { GenezioDeploy } from "@genezio/types"
+import mongoose from "mongoose";
+import { ObjectId } from "mongodb";
+import { TaskModel } from "./models/task";
+import { GenezioDeploy } from "@genezio/types";
+
+const red_color = "\x1b[31m%s\x1b[0m";
+const missing_env_error =
+  "ERROR: Your MONGO_DB_URI environment variable is not properly set, go to https://genez.io/blog/how-to-add-a-mongodb-to-your-genezio-project/ to learn how to integrate your project with Mongo DB";
 
 export type Task = {
-  id: string,
-  token: string,
-  title: string,
-  description: string,
-  solved: boolean,
-  date: string,
-}
+  id: string;
+  token: string;
+  title: string;
+  description: string;
+  solved: boolean;
+  date: string;
+};
 
 export type GetTasksResponse = {
-  success: boolean,
-  tasks: Task[]
-}
+  success: boolean;
+  tasks: Task[];
+  err: string;
+};
 
 export type GetTaskResponse = {
-  success: boolean,
-  task?: Task
-}
+  success: boolean;
+  task: Task;
+  err: string;
+};
 
 export type UpdateTaskResponse = {
-  success: boolean,
-}
+  success: boolean;
+  err: string;
+};
 
 export type DeleteTaskResponse = {
-  success: boolean,
-}
+  success: boolean;
+  err: string;
+};
 
 /**
  * The Task server class that will be deployed on the genezio infrastructure.
@@ -44,8 +51,14 @@ export class TaskService {
    * Private method used to connect to the DB.
    */
   #connect() {
-    mongoose.set('strictQuery', false);
-    mongoose.connect(MONGO_DB_URI);
+    if (!process.env.MONGO_DB_URI) {
+      console.log(red_color, missing_env_error);
+      return;
+    }
+    mongoose.connect(process.env.MONGO_DB_URI || "").catch((err) => {
+      console.log(err);
+      throw err;
+    });
   }
 
   /**
@@ -57,55 +70,72 @@ export class TaskService {
    * @returns An object containing two properties: { success: true, tasks: tasks }
    */
   async getAllTasksByUser(token: string): Promise<GetTasksResponse> {
-    console.log(`Get all tasks request received for user with token ${token}`)
+    if (!process.env.MONGO_DB_URI) {
+      console.log(red_color, missing_env_error);
+      return { success: false, tasks: [], err: missing_env_error };
+    }
+    console.log(`Get all tasks request received for user with token ${token}`);
 
-    const tasks: Task[] = (await TaskModel.find({token: token})).map((task) => {
-      return {
-        id: task._id.toString(),
-        token: task.token,
-        title: task.title,
-        description: task.description,
-        solved: task.solved,
-        date: task.date.toString()
-        }
-        });
-
-    if (tasks.length === 0) {
-      await TaskModel.create({
-        token: token,
-        title: "Check the other example projects",
-        description: "https://github.com/Genez-io/genezio-examples"
-      })
-
-      await TaskModel.create({
-        token: token,
-        title: "Check our documentation",
-        description: "https://docs.genez.io/genezio-documentation/"
-      })
-
-      await TaskModel.create({
-        token: token,
-        title: "Read our technical articles on genezio blog",
-        description: "https://genez.io/blog/"
-      })
-
-      const initTasks: Task[] = (await TaskModel.find({token: token})).map((task) => {
+    let tasks;
+    try {
+      tasks = (await TaskModel.find({ token: token })).map((task) => {
         return {
           id: task._id.toString(),
           token: task.token,
           title: task.title,
           description: task.description,
           solved: task.solved,
-          date: task.date.toString()
-          }
-          });
-
-      console.log(`Found ${initTasks.length} tasks for user with token ${token}`)
-      return { success: true, tasks: initTasks };
+          date: task.date.toString(),
+        };
+      });
+    } catch (error: any) {
+      return { success: false, tasks: [], err: error.toString() };
     }
 
-    console.log(`Found ${tasks.length} tasks for user with token ${token}`)
-    return { success: true, tasks: tasks };
+    if (tasks.length === 0) {
+      try {
+        await TaskModel.create({
+          token: token,
+          title: "Check the other example projects",
+          description: "https://github.com/Genez-io/genezio-examples",
+        });
+
+        await TaskModel.create({
+          token: token,
+          title: "Check our documentation",
+          description: "https://docs.genez.io/genezio-documentation/",
+        });
+
+        await TaskModel.create({
+          token: token,
+          title: "Read our technical articles on genezio blog",
+          description: "https://genez.io/blog/",
+        });
+
+        const initTasks: Task[] = (await TaskModel.find({ token: token })).map(
+          (task) => {
+            return {
+              id: task._id.toString(),
+              token: task.token,
+              title: task.title,
+              description: task.description,
+              solved: task.solved,
+              date: task.date.toString(),
+            };
+          }
+        );
+
+        console.log(
+          `Found ${initTasks.length} tasks for user with token ${token}`
+        );
+        return { success: true, tasks: initTasks, err: "" };
+      } catch (error: any) {
+        return { success: false, tasks: [], err: error.toString() };
+      }
+    }
+
+    console.log(`Found ${tasks.length} tasks for user with token ${token}`);
+    return { success: true, tasks: tasks, err: "" };
   }
 
   /**
@@ -118,23 +148,67 @@ export class TaskService {
    * @param {*} description The task's description.
    * @returns An object containing two properties: { success: true, tasks: tasks }
    */
-  async createTask(token: string, title: string, description: string) : Promise<GetTaskResponse> {
-    console.log(`Create task request received for user with title ${title}`)
+  async createTask(
+    token: string,
+    title: string,
+    description: string
+  ): Promise<GetTaskResponse> {
+    if (!process.env.MONGO_DB_URI) {
+      console.log(red_color, missing_env_error);
+
+      return {
+        success: false,
+        task: {
+          id: "err",
+          token: "err",
+          title: "err",
+          description: "err",
+          solved: false,
+          date: "",
+        },
+        err: missing_env_error,
+      };
+    }
+    console.log(`Create task request received for user with title ${title}`);
 
     if (!description) {
       description = "";
     }
 
-    const task = await TaskModel.create({
-      token: token,
-      title: title,
-      description: description,
-      solved: false,
-    });
+    let task;
+    try {
+      task = await TaskModel.create({
+        token: token,
+        title: title,
+        description: description,
+        solved: false,
+      });
+    } catch (error: any) {
+      return {
+        success: false,
+        task: {
+          id: "err",
+          token: "err",
+          title: "err",
+          description: "err",
+          solved: false,
+          date: "",
+        },
+        err: error.toString(),
+      };
+    }
 
     return {
       success: true,
-      task: { id: task._id.toString(), token: token, title: title, description: description, solved: false, date: ""}
+      task: {
+        id: task._id.toString(),
+        token: token,
+        title: title,
+        description: description,
+        solved: false,
+        date: "",
+      },
+      err: "",
     };
   }
 
@@ -150,26 +224,42 @@ export class TaskService {
    * @param {*} solved If the task is solved or not.
    * @returns An object containing two properties: { success: true }
    */
-  async updateTask(token: string, id: string , title: string, description: string, solved: boolean): Promise<UpdateTaskResponse>{
-    console.log(`Update task request received with id ${id} with title ${title} and solved value ${solved}`)
+  async updateTask(
+    token: string,
+    id: string,
+    title: string,
+    description: string,
+    solved: boolean
+  ): Promise<UpdateTaskResponse> {
+    if (!process.env.MONGO_DB_URI) {
+      console.log(red_color, missing_env_error);
+      return { success: false, err: missing_env_error };
+    }
+    console.log(
+      `Update task request received with id ${id} with title ${title} and solved value ${solved}`
+    );
 
     if (!description) {
       description = "";
     }
 
-    await TaskModel.updateOne(
-      {
-        _id: new ObjectId(id.toString()),
-        token: token
-      },
-      {
-        title: title,
-        description: description,
-        solved: solved
-      }
-    );
+    try {
+      await TaskModel.updateOne(
+        {
+          _id: new ObjectId(id.toString()),
+          token: token,
+        },
+        {
+          title: title,
+          description: description,
+          solved: solved,
+        }
+      );
+    } catch (error: any) {
+      return { success: false, err: error.toString() };
+    }
 
-    return { success: true };
+    return { success: true, err: "" };
   }
 
   /**
@@ -181,11 +271,19 @@ export class TaskService {
    * @param {*} id The task's id.
    * @returns An object containing one property: { success: true }
    */
-  async deleteTask(token:string, id: string): Promise<DeleteTaskResponse> {
-    console.log(`Delete task with id ${id} request received`)
+  async deleteTask(token: string, id: string): Promise<DeleteTaskResponse> {
+    if (!process.env.MONGO_DB_URI) {
+      console.log(red_color, missing_env_error);
+      return { success: false, err: missing_env_error };
+    }
+    console.log(`Delete task with id ${id} request received`);
 
-    await TaskModel.deleteOne({ _id: new ObjectId(id), token:token });
+    try {
+      await TaskModel.deleteOne({ _id: new ObjectId(id), token: token });
+    } catch (error: any) {
+      return { success: false, err: error.toString() };
+    }
 
-    return { success: true };
+    return { success: true, err: "" };
   }
 }
